Set up activity log module
