+++ moped-etl/prefect/tasks/activity_log.py
@@ -1,0 +1,35 @@
+import os
+
+import prefect
+from prefect import task
+
+# set up the prefect logging system
+logger = prefect.context.get("logger")
+
+
+@task
+def create_activity_log_sqs():
+    # Use boto3 to create SQS
+    logger.info("creating activity log SQS")
+    return True
+
+
+@task
+def create_activity_log_lambda():
+    # Use boto3 to create activity log event lambda
+    logger.info("creating activity log Lambda")
+    return True
+
+
+@task
+def remove_activity_log_sqs():
+    # Use boto3 to remove SQS
+    logger.info("removing activity log SQS")
+    return True
+
+
+@task
+def remove_activity_log_lambda():
+    # Use boto3 to remove activity log event lambda
+    logger.info("removing activity log Lambda")
+    return True
